requirements.txt
asgiref==3.7.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==2023.10
cryptography==39.0.1
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-debug-toolbar==4.2.0
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.9.0
urllib3==1.26.14
whitenoise==6.6.0
